Add a script to give the player infinite ammo
The display acts up because of how the ammo values are stored in
memory versus how they are displayed in the user interface. The ammo
value is separated into a chamber and reservoir values common to most
first person shooters.

Since the instruction responsible for subtracting the ammo is a shared
instruction this is the first step into figuring out how to give the
player infinite ammo. The result is that the user still has to reload
the chamber, and that the ammo reservoir counts up because the chamber
value is different but the total ammo value is not.

There are a few ways to correct for this.

* Figure out how to distinguish the shared subtraction code so that
the player and other entities are able to be distinguished.

This is not the best solution as the shared instruction is used as
part of the game loop, meaning that any changes here are more likely
to cause a crash.

* Modify the script so that instead of nopping out the subtract total
ammo instruction, to make it add the maximum value to the chamber
effectively causing a no-reload (which could also be configured as a
cheat option)

This is probably the better solution at this time. As from the tests I
ran, the total ammo instruction is only used by the player (makes sense
as the ammo instruction relates directly to the player) for all
weapons. The only issue is whether the address used can resolve the
chamber ammo address correctly.