Fixing Infinite Ammo AoB scan
Address referenced in the next instruction used to locate the AoB scan
changes on subsequent loads of the game. To overcome this these bytes
need to be changed to '??'. To make up for this, additional bytes may
need to be added to the end of the byte scan but so far it appears to
work, which makes the '??' bytes useless. But for some reason, Cheat
Engine keeps adding them - so maybe there is a case where a byte
matches which will make the scan return multiple results - preventing
the Infinite Ammo cheat from being activated.